pyproject.toml
[project]
name = "langgraph-workshop-notebooks"
version = "1.0.0"
description = "LangGraph workshop notebooks - Learn state management, ReAct agents, and production patterns"
readme = "README.md"
authors = [
    { name = "Calvin Remsburg", email = "calvin@cdot.io" }
]
requires-python = ">=3.11"
dependencies = [
    # Core LangGraph and LangChain
    "langgraph>=0.2.50",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    # Persistence (for Notebook 04)
    "langgraph-checkpoint-sqlite>=2.0.11",
    # Jupyter environment
    "jupyter>=1.1.1",
    "ipykernel>=6.30.1",
    "ipython>=9.5.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "typing_extensions>=4.8.0",
    # Optional: Browser automation (for web search examples)
    "playwright>=1.49.0",
    "pan-scm-sdk>=0.3.44",
    # CLI
    "typer>=0.20.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
# Development tools
dev = [
    "pytest>=7.4",
    "black>=24.3",
    "flake8>=7.1",
]

# Observability (optional for Notebook 04)
observability = [
    "langfuse>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["notebooks", "src"]

[project.scripts]
scm-agent = "src.cli.app:main"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", ".venv", "build", "dist"]
